requirements.txt
astroid==2.4.2
atomicwrites==1.4.0
attrs==19.3.0
colorama==0.4.3
docutils==0.16
iniconfig==1.0.1
isort==4.3.21
kivy-deps.glew==0.1.12
kivy-deps.gstreamer==0.1.18
kivy-deps.sdl2==0.1.23
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
pypiwin32==223
pytest==6.0.1
pywin32==228
six==1.15.0
toml==0.10.1
wrapt==1.12.1
XlsxWriter==1.3.1
